refactor: add `ILogger` test extensions
